html/browsers/windows/browsing-context-names/browsing-context-choose-self-2.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - the giving name is empty string</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/browsers.html#browsing-context-names">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="self2.html" style="display:none"></iframe>
<script>

var t1 = async_test("The current browsing context must be chosen if the giving name is empty string");

window.addEventListener("message", function (e) {
  t1.step(function () {
    assert_equals(e.data.name, "selfWin2", "The browsing context name should be 'selfWin2'.");
  });
  t1.done();
}, false);

</script>
